Modified receiver_cadmium.hpp acc. to coding conventions
